+++ xbmc/interfaces/legacy/Dialog.cpp
@@ -39,7 +39,7 @@ namespace XBMCAddon
   {
     static void XBMCWaitForThreadMessage(int message, int param1, int param2)
     {
-      ThreadMessage tMsg = {(DWORD)message, (DWORD)param1, (DWORD)param2};
+      ThreadMessage tMsg = {(DWORD)message, param1, param2};
       CApplicationMessenger::Get().SendMessage(tMsg, true);
     }
